Address the most egregious rpmlint warnings
`rpmlint` complains about a lot of things that don't matter, but
sometimes it's got a point. The 'release' field of the changelog
has been corrected, as has the 'provides' and permissions on the
udev rules configuration.

Additionally, a couple of files have been marked as %config files,
which should make upgrading a little cleaner in the future.